+++ apps/webapp/components/layout/session/login-dialog-content.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { BitcashAccessContentType } from '@/components/layout/session/login-dialog'
+import { BitcashAccessContentType } from '@/components/layout/session/session-dialog'
 import { Button } from '@/components/ui/button'
 import {
   DialogDescription,
@@ -11,7 +11,7 @@ import {
 import { useSession } from '@/hooks/use-session'
 import QRCode from 'react-qr-code'
 
-export function LoginDialogContent({
+export function SessionDialogContent({
   updateDialogContent
 }: {
   updateDialogContent: (dialog: BitcashAccessContentType) => void
